Feature/parameters count

+++ Dietician/Dietician/Models/ParameterResults.cs
@@ -2,14 +2,14 @@
 {
     public class ParameterResults
     {
-        public int Bmi { get; set; }
+        public double Bmi { get; set; }
         public string BmiLabel { get; set; }
-        public int Cmp { get; set; }
-        public int Protein { get; set; } 
-        public int Carbohydrate { get; set; }
-        public int Fat { get; set; }
-        public int Whr { get; set; }
+        public double Cmp { get; set; }
+        public double Protein { get; set; } 
+        public double Carbohydrate { get; set; }
+        public double Fat { get; set; }
+        public double Whr { get; set; }
         public string WhrLabel { get; set; }
-        public int FatLevel { get; set; }
+        public double FatLevel { get; set; }
     }
 }
